# openapi/payment-contract.openapi.yaml
openapi: 3.0.3
info:
  title: ECI Payment Service API (Team Contract)
  version: 1.0.0
  description: Integration contract aligned to implemented payment-service.
servers:
  - url: http://localhost:3004
paths:
  /health:
    get:
      summary: Health check
      responses:
        '200':
          description: Healthy
  /v1/payments:
    get:
      summary: List payments
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            default: 10
        - in: query
          name: status
          schema:
            type: string
        - in: query
          name: order_id
          schema:
            type: string
      responses:
        '200':
          description: Paginated list
  /v1/payments/charge:
    post:
      summary: Charge payment (idempotent)
      parameters:
        - in: header
          name: Idempotency-Key
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [order_id, amount, method]
              properties:
                order_id:
                  type: string
                amount:
                  type: number
                method:
                  type: string
      responses:
        '201':
          description: Charged
        '200':
          description: Idempotent replay
        '402':
          description: Failed
        '400':
          description: Validation error
  /v1/payments/{paymentId}/refund:
    post:
      summary: Refund payment (idempotent)
      parameters:
        - in: path
          name: paymentId
          required: true
          schema:
            type: string
        - in: header
          name: Idempotency-Key
          required: true
          schema:
            type: string
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                amount:
                  type: number
      responses:
        '201':
          description: Refunded
        '200':
          description: Idempotent replay
        '404':
          description: Not found
components:
  schemas:
    Error:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        correlationId:
          type: string